requirements.txt
annotated-types==0.6.0
backoff==2.2.1
black==23.11.0
certifi==2023.7.22
click==8.1.7
elastic-transport==8.10.0
elasticsearch==8.11.0
isort==5.12.0
loguru==0.7.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.0.3
pydantic_core==2.14.1
python-dotenv==1.0.0
typing_extensions==4.8.0
urllib3==2.1.0
